EDA
So far incomplete - I've managed to completely lose my previous file... not sure how. So will redo it asap, here's a start on genre. I don't think it will affect anyone given that we're not using genre, but the one-hot encoding in the cleaned file has errors. This file fixes them. Hopefully won't mess up anything for anyone but just an FYI.